Move all mini/max header state logic into the Header component to keep things decoupled
