+++ client/src/App.tsx
@@ -29,35 +29,9 @@ export default function App(): React.ReactElement {
 }
 
 const Root = (): React.ReactElement => {
-  const [isMiniHeader, setIsMiniHeader] = useState(false);
-  const headerValues = {
-    minHeight: 'h-24',
-    maxHeight: 'h-28',
-    minPaddingTop: 'mt-24',
-    maxPaddingTop: 'mt-28'
-  }
-
-  const handleScroll = (): void => {
-    const windowHeight = window.innerHeight;
-    const scrollValue = window.scrollY;
-    const isAtBottom = scrollValue >= windowHeight;
-
-    if (isAtBottom) {
-      setIsMiniHeader(true);
-    } else {
-      setIsMiniHeader(false);
-    }
-  };
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [handleScroll])
-
   return (
     <div className="w-full h-full min-h-screen flex flex-col items-center">
-      <Header isMiniHeader={isMiniHeader} setIsMiniHeader={setIsMiniHeader} headerValues={headerValues} />
+      <Header />
       <div className="w-full max-w-screen-2xl h-full min-h-fit mb-6 pt-28 flex grow justify-center items-center">
         <Outlet />
       </div>
